+++ src/cdk/v2/destinations/zoho/config.ts
@@ -1,0 +1,1 @@
+export const MAX_BATCH_SIZE: number = 100;