+++ src/cdk/v2/destinations/zoho/types.ts
@@ -1,0 +1,51 @@
+import {
+  RouterTransformationRequestData,
+  Destination,
+  Connection,
+  Metadata,
+} from '../../../../types';
+
+export type ZohoMetadata = Metadata & {
+  secret: {
+    accessToken: string;
+  };
+};
+
+export type TransformedResponseToBeBatched = {
+  upsertData: unknown[];
+  deletionData: string[];
+  upsertSuccessMetadata: ZohoMetadata[];
+  deletionSuccessMetadata: ZohoMetadata[];
+};
+
+export type DestConfig = {
+  trigger?: string;
+  addDefaultDuplicateCheck?: boolean;
+  multiSelectFieldLevelDecision: Array<{ from: string; to: string }>;
+  object: string;
+  identifierMappings: Array<{ from: string; to: string }>;
+};
+
+export type ConnectionConfig = {
+  destination: DestConfig;
+};
+
+export type Message = {
+  fields?: Record<string, unknown>;
+  action?: string;
+  identifiers?: Record<string, unknown>;
+};
+
+export type SearchRecordParams = {
+  identifiers: Record<string, unknown>;
+  metadata: { secret: { accessToken: string } };
+  destination: Destination;
+  destConfig: DestConfig;
+};
+
+export type ZohoRouterIORequest = RouterTransformationRequestData<
+  Message,
+  Destination,
+  Connection<ConnectionConfig>,
+  ZohoMetadata
+>;